requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
cachetools==5.5.1
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
Django==5.1.5
django-allauth==65.3.1
djangorestframework==3.15.2
gevent==24.11.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
pyparsing==3.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websocket==0.2.1
websocket-client==1.8.0
websockets==14.2
zope.event==5.0
zope.interface==7.2
